+++ src/components/organisms/controllPanel/presentation/fragments/edit/fragments/forms/DataRelationForm.tsx
@@ -5,10 +5,10 @@ import { DavitButton } from "../../../../../../../atomic/buttons/DavitButton";
 import { DavitCommentButton } from "../../../../../../../atomic/buttons/DavitCommentButton";
 import { DavitDeleteButton } from "../../../../../../../atomic/buttons/DavitDeleteButton";
 import { DavitDropDown } from "../../../../../../../atomic/dropdowns/DavitDropDown";
-import { Form } from "../../../../../../../common/fragments/forms/Form";
-import { FormBody } from "../../../../../../../common/fragments/forms/FormBody";
-import { FormFooter } from "../../../../../../../common/fragments/forms/FormFooter";
-import { FormHeader } from "../../../../../../../common/fragments/forms/FormHeader";
+import { Form } from "../../../../../../../atomic/forms/Form";
+import { FormBody } from "../../../../../../../atomic/forms/fragments/FormBody";
+import { FormFooter } from "../../../../../../../atomic/forms/fragments/FormFooter";
+import { FormHeader } from "../../../../../../../atomic/forms/fragments/FormHeader";
 import { useDataRelationViewModel } from "../viewmodels/DataRelationViewModel";
 import { FormDivider } from "./fragments/FormDivider";
 import { FormLabel, FormlabelAlign } from "./fragments/FormLabel";
