+++ src/components/organisms/controllPanel/presentation/fragments/edit/fragments/forms/DecisionForm.tsx
@@ -12,11 +12,11 @@ import { DecisionDropDown } from "../../../../../../../atomic/dropdowns/Decision
 import { GoToOptionDropDown } from "../../../../../../../atomic/dropdowns/GoToOptionDropDown";
 import { InstanceDropDown } from "../../../../../../../atomic/dropdowns/InstanceDropDown";
 import { StepDropDown } from "../../../../../../../atomic/dropdowns/StepDropDown";
+import { Form } from "../../../../../../../atomic/forms/Form";
+import { FormBody } from "../../../../../../../atomic/forms/fragments/FormBody";
+import { FormFooter } from "../../../../../../../atomic/forms/fragments/FormFooter";
+import { FormHeader } from "../../../../../../../atomic/forms/fragments/FormHeader";
 import { DavitTextInput } from "../../../../../../../atomic/textinput/DavitTextInput";
-import { Form } from "../../../../../../../common/fragments/forms/Form";
-import { FormBody } from "../../../../../../../common/fragments/forms/FormBody";
-import { FormFooter } from "../../../../../../../common/fragments/forms/FormFooter";
-import { FormHeader } from "../../../../../../../common/fragments/forms/FormHeader";
 import { useDecisionViewModel } from "../viewmodels/DecisionViewModel";
 import { FormDivider } from "./fragments/FormDivider";
 import { FormLabel, FormlabelAlign } from "./fragments/FormLabel";
